requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
newsapi-python==0.0.2
requests==2.13.0
Werkzeug==0.14.1
